// server/types.ts
export interface EcosystemToken {
  id: number
  name: string
  currency: string
  chain: string
  network?: string
  type: string
  contract: string
  decimals: number
  status?: boolean
  precision?: EcosystemPrecision
  limits?: EcosystemTokenLimits
  fees?: any
  contractType?: EcosystemTokenContractType
  created_at: Date
}

export enum EcosystemTokenContractType {
  PERMIT = 'PERMIT',
  NO_PERMIT = 'NO_PERMIT',
  NATIVE = 'NATIVE',
}

export type EcosystemPrecision = {
  amount: number
  price: number
}

export type EcosystemTokenLimits = {
  deposit: {
    min: number
    max: number
  }
  withdraw: {
    min: number
    max: number
  }
}